backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
sqlalchemy==2.0.31
pymysql==1.1.1
python-dotenv==1.0.1
passlib==1.7.4
flask_sqlalchemy==3.1.1
